move the network plugin
